feat: direction( 경로 ) 마이페이지 Controller 공용으로 경로 수정 및 HTML 이동 경로 수정
