+++ service/src/main/java/com/petkpetk/service/common/controller/seller/SellerMyPageController.java
@@ -1,4 +1,4 @@
-package com.petkpetk.service.domain.shopping.controller;
+package com.petkpetk.service.common.controller.seller;
 
 import java.util.Optional;
 
@@ -23,13 +23,14 @@ import lombok.RequiredArgsConstructor;
 
 @Controller
 @RequiredArgsConstructor
-@RequestMapping("/seller")
-public class SellerItemController {
+@RequestMapping("/seller/my-page/")
+public class SellerMyPageController {
 	private final UserAccountService userAccountService;
 	private final ItemService itemService;
 
 	@GetMapping("/information")
-	public String informationView(Model model, @AuthenticationPrincipal UserAccountPrincipal userAccountPrincipal) {
+	public String informationView(
+		Model model, @AuthenticationPrincipal UserAccountPrincipal userAccountPrincipal) {
 		model.addAttribute("userAccount", userAccountService.getUserUpdateRequestView(userAccountPrincipal));
 		ProfileImage profileImage = userAccountService.getUserProfile(userAccountPrincipal);
 		model.addAttribute("profileImage", profileImage);
@@ -37,7 +38,8 @@ public class SellerItemController {
 	}
 
 	@GetMapping(value = {"/item-manage", "/item-manage/{page}"})
-	public String itemManage(ItemSearchDto itemSearchDto, @PathVariable("page") Optional<Integer> page, Model model, Authentication authentication){
+	public String itemManage(
+		ItemSearchDto itemSearchDto, @PathVariable("page") Optional<Integer> page, Model model, Authentication authentication){
 		String email = authentication.getName();
 
 		PageRequest pageRequest = PageRequest.of(page.orElse(0), 5);
