+++ service/src/main/resources/templates/fragments/header.html
@@ -41,16 +41,16 @@
                 <ul class="flex justifyContent topList">
                     <li class="log"><a th:href="@{/logout}">로그아웃</a></li>
                     <li >
-                            <a th:href="@{/user/information}">
+                            <a th:href="@{/user/my-page/information}">
                                 <div class="profile Pone" onmouseover="showMenu()" onmouseleave="hideMenu()">
                                     <img class="userProfile" th:src="@{/images/pinkMoala.png}">
                                     <div id="topHoverMenu" class="topHoverMenu">
                                         <img class="ladder" th:src="@{/images/ladder.png}">
                                         <ul class="menuBox">
                                             <li class="nickName" th:text="${#authentication.getPrincipal().getNickname()}">내 닉네임</li>
-                                            <li class="menus"><a th:href="@{/user/information}">마이 페이지</a></li>
+                                            <li class="menus"><a th:href="@{/user/my-page/information}">마이 페이지</a></li>
                                             <li class="menus"><a th:href="@{/cart}">장바구니</a></li>
-                                            <li class="menus"><a th:href="@{/order}">구매 이력</a></li>
+                                            <li class="menus"><a th:href="@{/user/order/orders}">구매 이력</a></li>
                                             <li class="menus"><a th:href="@{/QnA}">QnA</a></li>
                                             <li class="menus"><a th:href="@{/logout}">로그아웃</a></li>
                                         </ul>
@@ -64,15 +64,15 @@
                 <ul class="flex justifyContent topList">
                     <li class="log"><a th:href="@{/logout}">로그아웃</a></li>
                     <li >
-                            <a  th:href="@{/seller/information}">
+                            <a  th:href="@{/seller/my-page/information}">
                                 <div class="profile Pone" onmouseover="showMenu()" onmouseleave="hideMenu()">
                                     <img class="userProfile" th:src="@{/images/pinkMoala.png}">
                                     <div id="topHoverMenu" class="topHoverMenu">
                                         <img class="ladder" th:src="@{/images/ladder.png}">
                                         <ul class="menuBox">
                                             <li class="nickName" th:text="${#authentication.getPrincipal().getNickname()}">내 닉네임</li>
-                                            <li class="menus"><a th:href="@{/seller/information}">사업 정보</a></li>
-                                            <li class="menus"><a th:href="@{/seller/item-manage}">상품 관리</a></li>
+                                            <li class="menus"><a th:href="@{/seller/my-page/information}">사업 정보</a></li>
+                                            <li class="menus"><a th:href="@{/seller/my-page/item-manage}">상품 관리</a></li>
                                             <li class="menus"><a>주문 상태</a></li>
                                             <li class="menus"><a th:href="@{/QnA}">QnA</a></li>
                                             <li class="menus"><a th:href="@{/logout}">로그아웃</a></li>
@@ -121,12 +121,12 @@
                 <a><li>QnA</li></a>
                 <div class="line"></div>
                 <!------------------------------ if user == null ------------------------------>
-                <a th:href="@{/seller/information}"><li sec:authorize="hasAnyAuthority('ROLE_SELLER')">사업 정보 <span>My Page</span></li></a>
-                <a th:href="@{/seller/item-manage}"><li sec:authorize="hasAnyAuthority('ROLE_SELLER')">상품 관리 <span>Manage Products</span></li></a>
+                <a th:href="@{/seller/my-page/information}"><li sec:authorize="hasAnyAuthority('ROLE_SELLER')">사업 정보 <span>My Page</span></li></a>
+                <a th:href="@{/seller/my-page/item-manage}"><li sec:authorize="hasAnyAuthority('ROLE_SELLER')">상품 관리 <span>Manage Products</span></li></a>
                 <a><li sec:authorize="hasAnyAuthority('ROLE_SELLER')">주문 관리 <span>Manage Orders</span></li></a>
                 <a><li sec:authorize="hasAnyAuthority('ROLE_SELLER')">정산 <span>Calculate</span></li></a>
 
-                <a th:href="@{/user/information}"><li sec:authorize="hasAnyAuthority('ROLE_USER')">마이 페이지 <span>My Page</span></li></a>
+                <a th:href="@{/user/my-page/information}"><li sec:authorize="hasAnyAuthority('ROLE_USER')">마이 페이지 <span>My Page</span></li></a>
                 <a th:href="@{/order}"><li sec:authorize="hasAnyAuthority('ROLE_USER')">구매 이력 <span>purchase history</span></li></a>
                 <a th:href="@{/cart}"><li sec:authorize="!hasAnyAuthority('ROLE_ADMIN', 'ROLE_SELLER')">장바구니 <span>Cart</span></li></a>
