+++ service/src/main/resources/templates/fragments/myPageMenubar.html
@@ -6,8 +6,8 @@
 
     <div sec:authorize="hasAnyAuthority('ROLE_SELLER')">
         <ul class="sellerMenubar bar">
-            <li><a th:href="@{/seller/information}">사업 정보</a></li>
-            <li><a th:href="@{/seller/item-manage}">등록 상품</a></li>
+            <li><a th:href="@{/seller/my-page/information}">사업 정보</a></li>
+            <li><a th:href="@{/seller/my-page/item-manage}">등록 상품</a></li>
             <li><a>주문 내역</a></li>
             <li><a>정산</a></li>
         </ul>
@@ -15,10 +15,10 @@
 
     <div sec:authorize="hasAnyAuthority('ROLE_USER')">
         <ul class="userMenubar bar">
-            <li><a th:href="@{/user/information}">마이 페이지</a></li>
+            <li><a th:href="@{/user/my-page/information}">마이 페이지</a></li>
             <li><a>장바구니</a></li>
-            <li><a>구매 내역</a></li>
-            <li><a th:href="@{/user/reviewHistory}">리뷰 내역</a></li>
+            <li><a th:href="@{/user/order/orders}">구매 내역</a></li>
+            <li><a th:href="@{/user/my-page/reviewHistory}">리뷰 내역</a></li>
             <li><a>커뮤니티 활동</a></li>
         </ul>
     </div>
